chore: remove unimplemented fromStupidString
There are no longer any references to fromStupidString in the
multiaddr and go-multiaddr repos, and it's always been a needlesly
inflamtory method name.

This PR remove is and the test that checked it was unimplemented.

There are efforts to provide encoding and decoding between
multiaddrs and uris in

- https://github.com/tableflip/multiaddr-to-uri
- https://github.com/tableflip/uri-to-multiaddr

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>