fix: return correct family from toOptions with dns4 MultiAddr (#233)
When calling toOptions with a dns4 multiaddress, a MultiaddrObject is returned where the ip family is 6 instead of 4. This causes an issue where libp2p is unable to resolve the domain and dial the peer. This only appears to be an issue on some systems. It works on Mac OS but fails on for example Alpine Linux.

Example:
```javascript
const {multiaddr} = require("multiaddr")

console.log(multiaddr('/dns4/google.net/tcp/8000').toOptions())`
```

```javascript
Object {family: 6, host: "google.net", transport: "tcp", port: 8000}
```

Co-authored-by: achingbrain <alex@achingbrain.net>