fix: fix converting strings to IP addresses and back again
Partial fix for the issues observed in the ipfs interop suite - windows
seems to lose it's reference to the `results` variable during node
startup, but only windows.  Instead return the result from the scope
in which is is created.

A follow up will be needed to address the other windows problem which
is the IP detection regex seems to stop working.  Might be a bug in
node I guess?