feat!: convert to interface
- Renames the `Multiaddr` class to `DefaultMultiaddr`
- Exports `Multiaddr` as an interface

The `multiaddr` function should be used to create new instances of
`Multiaddr` (introduced in v9.0.0) instead of instantiating the
`Multiaddr` class directly.

This allows us to remove the `inspect` symbol from the interface
and avoid use of `Object.defineProperty` which is very slow.

BREAKING CHANGE: the `Multiaddr` class is now an interface