fix: make getPeerId return the target peer id from relay addresses
If a p2p-circuit address doesn't contain the `/p2p/QmFoo` tuple for
the target peer, `getPeerId` returns the peer id of the relay which
is not what you'd expect.

The fix here is for relay addresses, to return the peer id of the
target and never the relay.

Fixes #319